requirements.txt
openai
chromadb
langdetect
pypdf
python-dotenv
streamlit
tiktoken
gdown
